# contracts/http/openapi.yaml
...
openapi: 3.1.0
info:
  title: Research Radar — HTTP wire contract
  version: 0.3.0
  summary: 'Hợp đồng HTTP cho toàn bộ operation `transport: http` của contracts/ports.yaml.'
  description: |-
    Đây là hợp đồng dây (wire) của Research Radar, sinh từ `contracts/ports.yaml` và giữ nguyên `operation_id`
    làm `operationId`. Mỗi operation `transport: http` trong ports.yaml có ĐÚNG MỘT cặp path+method ở đây.

    Những gì file này KHÔNG định nghĩa, và nơi chúng thuộc về:
      - Ngữ nghĩa transition trạng thái  → contracts/state/*.yaml (PC03)
      - Danh mục mã lỗi và envelope      → contracts/errors.yaml (PC03); `ErrorEnvelope` dưới đây là bản sao ĐÚNG
                                           hình dạng của `error_envelope` trong file đó
      - Ngân sách retry/timeout/lease    → contracts/retry-policy.yaml (PC03)
      - Ranh giới module và cạnh bị cấm  → contracts/modules.yaml (PC01)
      - Nội dung/coverage báo cáo        → contracts/reporting/* (PC04)
      - Task và schema AI                → contracts/ai/* (PC06)
      - Ngữ nghĩa lệnh Telegram          → contracts/telegram/commands.yaml (PC07)
      - Vòng đời secret, CSRF, xoay token→ contracts/ops/secrets.md (PC08)

    Wire rules bắt buộc (SRC-PLAN §5.1), áp cho mọi operation:
      - ID truyền dưới dạng string; timestamp UTC RFC 3339 với độ chính xác mili giây (AMD-B08).
      - Mutation từ worker mang `request_id`, `schema_version`, `idempotency_key`, `payload_hash`; mutation đang
...

...
  x-contract:
...
    - REQ-D04
...
servers:
...
security:
- ownerSessionCookie: []
  ownerCsrfToken: []
components:
  securitySchemes:
    ownerSessionCookie:
      type: apiKey
      in: cookie
      name: rr_session
      description: 'Cookie phiên HttpOnly + Secure + SameSite=Lax. PROVISIONAL, phối hợp với PC08 (xem info.x-contract.security_scheme_decision).
        Với MUTATION, scheme này xuất hiện CÙNG `ownerCsrfToken` trong MỘT security requirement — tức quan hệ AND, cookie
        một mình KHÔNG đủ. Với read-only, cookie đứng một mình: CSRF bảo vệ thao tác đổi trạng thái, không phải phép đọc.'
    ownerCsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 'CSRF double-submit: giá trị phải khớp cookie `rr_csrf` (không-HttpOnly). Bắt buộc cho MỌI mutation của
        owner session. Thiếu hoặc lệch ⇒ 403 `CSRF_REJECTED` (KHÔNG phải `FORBIDDEN_EDGE`: cạnh vẫn hợp lệ, chỉ lời gọi
        không chứng minh được nó do trang của ứng dụng phát ra — ruling CR-PC08-03/CR-PC08-04). Phiên KHÔNG bị hủy khi gặp
        mã này.'
    collectorToken:
      type: http
      scheme: bearer
      description: Token của collector, lưu trong file cấu hình QUYỀN HẠN CHẾ trên máy cá nhân, KHÔNG BAO GIỜ trong repo
        (SRC-SPEC §11.2). Chỉ mở được các operation `worker.*` và `ingest.*`; gọi `save.create` bằng token này ⇒ 401 (denied
        case NC-01).
    analysisWorkerToken:
      type: http
      scheme: bearer
      description: Token của analysis worker. Chỉ mở `analysis.*` và `secret.issue_task_credential`. KHÔNG mở đường tới
        tag, report hay Saved (denied case NC-02).
    telegramIngressSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Bí mật webhook do app đặt khi đăng ký webhook với Telegram. Sai hoặc thiếu ⇒ 401 và KHÔNG xử lý update.
        Biết chat ID KHÔNG phải là quyền (SRC-SPEC §11.3).
    backupOperatorToken:
      type: http
      scheme: bearer
      description: 'Danh tính operator backup. Tách khỏi phiên owner vì restore là thao tác có khóa side effect (AMD-B11).
        Sửa ở FIX1 (CR-PC08-02): mọi route `backup.*` nhận DUY NHẤT scheme này — trước đây phiên trình duyệt cũng qua được,
        nghĩa là một tab đang mở có thể kích hoạt restore.'
...
paths:
...
  /v1/runs:
    get:
      operationId: run.list
      summary: Danh sách run + trạng thái collector online/offline
      description: 'Bộ lọc phân trang; với caller Telegram chỉ lấy N run gần nhất.


        Không. Lệnh status **không** mutation (B10).


        Lệnh Telegram `status` là READ-ONLY và đi qua đây (B10/AMD-B10). Mỗi run trả bộ ba `(status, outcome, stop_reason)`
        để ba trạng thái của REQ-AC15 phân biệt được (I13).'
      tags:
      - runs
      security:
      - ownerSessionCookie: []
      - telegramIngressSecret: []
      x-auth-scope: owner_session_or_linked_chat
      x-owner-module: MOD-job-service
      x-caller-modules:
      - MOD-web-ui
      - MOD-telegram-adapter
      x-mutation: false
      x-scenario-refs:
      - SC15
      x-requirement-refs:
      - REQ-P0-12
      - REQ-D11
      - REQ-D36
      - REQ-AC15
      x-decision-refs:
      - B02
      - B10
      - AMD-B02
      x-invariant-refs:
      - I01
      - I13
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/PageLimit'
        description: Số run tối đa.
      - name: cursor
        in: query
        required: false
        schema:
          type: string
        description: Con trỏ trang.
      - $ref: '#/components/parameters/SchemaVersionHeader'
      - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          description: 'Mỗi run: phase, status, outcome, stop_reason, trigger_type, tiến độ. Ba trạng thái `empty`, `stopped/limit`,
            `failed` hiển thị khác nhau (SRC-SPEC §8.3, I13).'
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/GenericObject'
                x-schema-source: contracts/state/run.yaml
                description: Hình dạng do hợp đồng `contracts/state/run.yaml` sở hữu (không phải JSON Schema nên không $ref
                  được). Gói sở hữu phải thay bằng schema đóng trước G3.
          headers:
            X-Schema-Version:
              description: Phiên bản schema của response.
              schema:
                $ref: '#/components/schemas/SemVer'
        '401':
          description: 'Lỗi: `UNAUTHORIZED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 'Lỗi: `UNAUTHORIZED_COMMAND`, `FORBIDDEN_EDGE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: 'Lỗi: `INTERNAL`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/runs/{run_id}:
    get:
      operationId: run.get
      summary: Chi tiết một run (Run detail)
      description: 'run_id.


        Log trả về ĐÃ CHE SECRETS (SRC-SPEC §11.2).'
      tags:
      - runs
      security:
      - ownerSessionCookie: []
      x-auth-scope: owner_session
      x-owner-module: MOD-job-service
      x-caller-modules:
      - MOD-web-ui
      x-mutation: false
      x-scenario-refs:
      - SC03
      - SC15
      x-requirement-refs:
      - REQ-P0-12
      - REQ-D14
      x-decision-refs:
      - B02
      x-invariant-refs:
      - I01
      - I11
      - I13
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Ulid'
        description: Run cần xem.
      - $ref: '#/components/parameters/SchemaVersionHeader'
      - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          description: trigger_type, cấu hình đã áp, số bài lấy, số bài mới, checkpoint đã ACK, lỗi, log **đã che secrets**.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/GenericObject'
                x-schema-source: contracts/state/run.yaml
                description: Hình dạng do hợp đồng `contracts/state/run.yaml` sở hữu (không phải JSON Schema nên không $ref
                  được). Gói sở hữu phải thay bằng schema đóng trước G3.
          headers:
            X-Schema-Version:
              description: Phiên bản schema của response.
              schema:
                $ref: '#/components/schemas/SemVer'
        '401':
          description: 'Lỗi: `UNAUTHORIZED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 'Lỗi: `FORBIDDEN_EDGE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 'Lỗi: `NOT_FOUND`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: 'Lỗi: `INTERNAL`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/runs/run-now:
    post:
      operationId: run.run_now
      summary: Chạy ngay (thủ công)
      description: 'Không có tham số nghiệp vụ ngoài request_id.


        Tạo run `trigger_type = manual` ở `queued`. **Không** vượt qua run đang `needs_user` hay `blocked` (B10): trả run
        hiện tại kèm lý do, không tự resume.


        KHÔNG vượt qua run đang `needs_user` hay `blocked`: trả run hiện tại kèm lý do, KHÔNG tự resume (B10/AMD-B10, contracts/state/run.yaml
        T-RUN-21).'
      tags:
      - runs
      security:
      - ownerSessionCookie: []
        ownerCsrfToken: []
      - telegramIngressSecret: []
      x-auth-scope: owner_session_or_linked_chat
      x-owner-module: MOD-job-service
      x-caller-modules:
      - MOD-web-ui
      - MOD-telegram-adapter
      x-mutation: true
      x-scenario-refs:
      - SC04
      - SC15
      x-requirement-refs:
      - REQ-P0-10
      - REQ-D10
      - REQ-D14
      - REQ-D36
      x-decision-refs:
      - B10
      - AMD-B10
      x-invariant-refs:
      - I01
      - I10
      - I13
      x-idempotency:
        required: true
        key_scope: request_id
        rule_vi: Cùng request_id không tạo hai run; bấm lặp trả run đang có.
        header: Idempotency-Key
      parameters:
      - $ref: '#/components/parameters/SchemaVersionHeader'
      - $ref: '#/components/parameters/RequestIdHeader'
      - $ref: '#/components/parameters/IdempotencyKeyHeader'
      - $ref: '#/components/parameters/CsrfTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
      responses:
        '200':
          description: Trả run đã xếp hàng, HOẶC run đang chạy hiện có (chính sách COALESCE).
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/GenericObject'
                x-schema-source: contracts/state/run.yaml
                description: Hình dạng do hợp đồng `contracts/state/run.yaml` sở hữu (không phải JSON Schema nên không $ref
                  được). Gói sở hữu phải thay bằng schema đóng trước G3.
          headers:
            X-Schema-Version:
              description: Phiên bản schema của response.
              schema:
                $ref: '#/components/schemas/SemVer'
        '401':
          description: 'Lỗi: `UNAUTHORIZED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 'Lỗi: `CSRF_REJECTED`, `UNAUTHORIZED_COMMAND`, `FORBIDDEN_EDGE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 'Lỗi: `IDEMPOTENCY_CONFLICT`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 'Lỗi: `VALIDATION_ERROR`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 'Lỗi: `RATE_LIMITED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            Retry-After:
              description: Giây do nguồn/API chỉ định. Giá trị này THẮNG mọi backoff cấu hình.
              schema:
                type: integer
                minimum: 0
        '500':
          description: 'Lỗi: `INTERNAL`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 'Lỗi: `STORAGE_WRITE_FAILED`. `storage.health = write_blocked` hoặc `maintenance`: mutation KHÔNG
            được ACK và KHÔNG có gì được ghi (SRC-PLAN §8.4, I02).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            Retry-After:
              description: Giây. Gợi ý theo `storage_recovery_probe` (30 s).
              schema:
                type: integer
                minimum: 0
...
  /v1/runs/{run_id}/cancel:
    post:
      operationId: run.cancel
      summary: Hủy một run chưa kết thúc
      description: 'run_id.


        Thu hồi lease, chặn publish mới; dữ liệu đã commit không bị xóa (SRC-PLAN §8.1).


        Thu hồi lease, chặn publish mới; dữ liệu ĐÃ COMMIT không bị xóa.'
      tags:
      - runs
      security:
      - ownerSessionCookie: []
        ownerCsrfToken: []
      x-auth-scope: owner_session
      x-owner-module: MOD-job-service
      x-caller-modules:
      - MOD-web-ui
      x-mutation: true
      x-scenario-refs:
      - SC20
      x-requirement-refs:
      - REQ-P0-12
      x-decision-refs:
      - B02
      x-invariant-refs:
      - I10
      x-idempotency:
        required: true
        key_scope: request_id
        rule_vi: Hủy lại run đã cancelled trả kết quả idempotent.
        header: Idempotency-Key
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Ulid'
        description: Run chưa kết thúc.
      - $ref: '#/components/parameters/SchemaVersionHeader'
      - $ref: '#/components/parameters/RequestIdHeader'
      - $ref: '#/components/parameters/IdempotencyKeyHeader'
      - $ref: '#/components/parameters/CsrfTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
      responses:
        '200':
          description: Trạng thái run sau khi hủy.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/GenericObject'
                x-schema-source: contracts/state/run.yaml
                description: Hình dạng do hợp đồng `contracts/state/run.yaml` sở hữu (không phải JSON Schema nên không $ref
                  được). Gói sở hữu phải thay bằng schema đóng trước G3.
          headers:
            X-Schema-Version:
              description: Phiên bản schema của response.
              schema:
                $ref: '#/components/schemas/SemVer'
        '401':
          description: 'Lỗi: `UNAUTHORIZED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 'Lỗi: `CSRF_REJECTED`, `FORBIDDEN_EDGE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 'Lỗi: `NOT_FOUND`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 'Lỗi: `IDEMPOTENCY_CONFLICT`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 'Lỗi: `VALIDATION_ERROR`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 'Lỗi: `RATE_LIMITED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            Retry-After:
              description: Giây do nguồn/API chỉ định. Giá trị này THẮNG mọi backoff cấu hình.
              schema:
                type: integer
                minimum: 0
        '500':
          description: 'Lỗi: `INTERNAL`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...